+++ packages/app/src/import/service/importer.service.ts
@@ -8,7 +8,8 @@ import {
     CategoryEntityInterface,
     ExternalSourceEnum,
     InstrumentEntityInterface,
-    TransactionCreateEntityInterface,
+    TransactionCreateInputInterface,
+    TransactionEntryCreateInputInterface,
     TransactionEntryTypeEnum,
     TransactionTypeEnum,
     UserIconNameEnum
@@ -117,8 +118,8 @@ export class ImporterService {
         };
     }
 
-    private async processTransactions(csvText: string, progress: ImportProgressInterface): Promise<TransactionCreateEntityInterface[]> {
-        const transactions: TransactionCreateEntityInterface[] = [];
+    private async processTransactions(csvText: string, progress: ImportProgressInterface): Promise<TransactionCreateInputInterface[]> {
+        const transactions: TransactionCreateInputInterface[] = [];
 
         await this.processRows(csvText, (normalizedRow, row) => {
             progress.processed += 1;
@@ -136,7 +137,7 @@ export class ImporterService {
         return transactions;
     }
 
-    private createTransaction(normalizedRow: NormalizedRow): TransactionCreateEntityInterface {
+    private createTransaction(normalizedRow: NormalizedRow): TransactionCreateInputInterface {
         const { toAccount, fromAccount, category, operatedAt, toAmount, fromInstrument, toInstrument, fromAmount } =
             this.parseRow(normalizedRow);
 
@@ -191,13 +192,12 @@ export class ImporterService {
         category: CategoryEntityInterface,
         source: EntryParams,
         dest: EntryParams | null
-    ): TransactionCreateEntityInterface['entries'] {
+    ): TransactionEntryCreateInputInterface[] {
         if (type === TransactionTypeEnum.INCOME) {
             return [
                 {
                     type: TransactionEntryTypeEnum.CREDIT,
                     amount: Math.abs(source.amount),
-                    instrumentId: source.instrument.id,
                     accountId: source.account.id,
                     categoryId: category.id
                 }
@@ -207,7 +207,6 @@ export class ImporterService {
                 {
                     type: TransactionEntryTypeEnum.DEBIT,
                     amount: Math.abs(source.amount),
-                    instrumentId: source.instrument.id,
                     accountId: source.account.id,
                     categoryId: category.id
                 }
@@ -217,14 +216,12 @@ export class ImporterService {
                 {
                     type: TransactionEntryTypeEnum.CREDIT,
                     amount: Math.abs(source.amount),
-                    instrumentId: source.instrument.id,
                     accountId: source.account.id,
                     categoryId: category.id
                 },
                 {
                     type: TransactionEntryTypeEnum.DEBIT,
                     amount: Math.abs(dest.amount),
-                    instrumentId: dest.instrument.id,
                     accountId: dest.account.id,
                     categoryId: category.id
                 }